+++ examples/example-auth.php
@@ -11,7 +11,19 @@ else
     die("Please rename the config-sample.php file to config.php and add your App.net client id and secret to it\n");
 }
 
-require_once dirname(__FILE__) . '/DPZAppNet.php';
+spl_autoload_register(function($className)
+{
+    $className = str_replace ('\\', DIRECTORY_SEPARATOR, $className);
+    include (dirname(__FILE__) . '/../src/' . $className . '.php');
+});
+
+/**
+ * Or use this:
+ * `./composer.phar install`
+ * require_once dirname(__DIR__) . '/vendor/autoload.php';
+ */
+
+use \DPZ\AppNet;
 
 // Build the URL for our callback
 $callback = sprintf('%s://%s%s%s/example-auth.php',
@@ -21,7 +33,7 @@ $callback = sprintf('%s://%s%s%s/example-auth.php',
     $appNetRedirectPathPrefix
     );
 
-$appNet = new DPZFAppNet($appNetClientId, $appNetClientSecret, $callback);
+$appNet = new AppNet($appNetClientId, $appNetClientSecret, $callback);
 
 if (!$appNet->authenticate('stream write_post'))
 {