+++ src/DPZ/AppNet.php
@@ -21,11 +21,11 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  *
  */
-namespace DPZF;
+namespace DPZ;
 
 class AppNet
 {
-    const VERSION = 0.1;
+    const VERSION = 0.2;
 
     /**
      * Session variable name used to store authentication data