+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-Version 1.0 - 13th August 2012
+Version 0.2 - 16th August 2012
+==============================
+
+Moved class into DPZ namespace
+Added composer.json
+
+Thanks to Till Klampaeckel for initiating the namespace move and providing the composer.json
+
+Version 0.1 - 13th August 2012
 ==============================
 
 Initial version.